// src/pages/modules/SPI/external-portal/data/types.ts
export interface IExternalPortal {
  id_portal_eksternal: string
  id_satuan_organisasi: string
  url_gambar: string
  key_gambar: string
  url: string
  urutan: number
  created_at: string
  created_user: string
  updated_at: string
  updated_user: string
  nama_user_created: string
  nama_user_updated: string
}